Switched from unordered map to vector for block data
unordered map was a massive performance drawback as multiple memory reashing caused it to have terrible locality. array<T, N> from C++11 was then tried and had incredible locality m,but it was discovered that these objects are stored in the stack rather than in the heap hence 1000s of blocks were bound to causing a stack overflow. vectors proved to have better memory locality than unordered maps but are stored in the heap. Best world from both.